chore: improve TS types, refactor code
