+++ components/image-preview/utils/loadStatus.ts
@@ -1,0 +1,8 @@
+enum LOAD_STATUS {
+  before = 'loadBefore',
+  start = 'loadStart',
+  end = 'loadEnd',
+  after = 'loadAfter',
+}
+
+export default LOAD_STATUS;
